split config cmd to files
