+++ cmd/kafeman/config_cmd/set_cluster.go
@@ -5,68 +5,12 @@ import (
 	"os"
 	"strings"
 
-	"github.com/worldbug/kafeman/cmd/kafeman/common"
-	"github.com/worldbug/kafeman/cmd/kafeman/completion_cmd"
-	"github.com/worldbug/kafeman/cmd/kafeman/run_configuration"
-	"github.com/worldbug/kafeman/internal/config"
-	"github.com/worldbug/kafeman/internal/logger"
-
 	"github.com/manifoldco/promptui"
 	"github.com/spf13/cobra"
+	"github.com/worldbug/kafeman/cmd/kafeman/completion_cmd"
+	"github.com/worldbug/kafeman/cmd/kafeman/run_configuration"
+	"github.com/worldbug/kafeman/internal/logger"
 )
-
-func newConfigOptions() *configOptions {
-	return &configOptions{}
-}
-
-type configOptions struct{}
-
-// TODO: refactor
-func NewConfigCMD() *cobra.Command {
-	// if config not inited
-	if len(run_configuration.Config.Clusters) == 0 {
-		run_configuration.SetCurrentCluster("local")
-		run_configuration.SetCluster(
-			config.Cluster{
-				Name:    "local",
-				Brokers: []string{"localhost:9092"},
-			},
-		)
-	}
-
-	cmd := &cobra.Command{
-		Use:   "config",
-		Short: "Handle kafman configuration",
-	}
-
-	// TODO:
-	configPath := ""
-	// ConfigCMD.AddCommand(configImportCmd)
-	// ConfigCMD.AddCommand(configUseCmd)
-	// ConfigCMD.AddCommand(configLsCmd)
-	// ConfigCMD.AddCommand(configAddClusterCmd)
-	// ConfigCMD.AddCommand(configRemoveClusterCmd)
-	cmd.AddCommand(NewConfigSetCluster(configPath))
-	cmd.AddCommand(NewConfigCurrentContextCMD(configPath))
-	// ConfigCMD.AddCommand(configAddEventhub)
-
-	cmd.AddCommand(NewConfigInitCMD(configPath))
-
-	return cmd
-}
-
-func NewConfigCurrentContextCMD(configPath string) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "current-context",
-		Short: "Displays the current context",
-		Args:  cobra.ExactArgs(0),
-		Run: func(cmd *cobra.Command, args []string) {
-			fmt.Fprintf(os.Stdout, "%s\n", run_configuration.GetCurrentCluster().Name)
-		},
-	}
-
-	return cmd
-}
 
 func NewConfigSetCluster(configPath string) *cobra.Command {
 	cmd := &cobra.Command{
@@ -131,24 +75,6 @@ func NewConfigSetCluster(configPath string) *cobra.Command {
 				fmt.Fprintf(os.Stderr, "Can`t save config: %+v", err)
 				os.Exit(0)
 			}
-		},
-	}
-
-	return cmd
-}
-
-func NewConfigInitCMD(configPath string) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "init",
-		Short: "Create empty config and export to file (default ~/.kafeman/config.yaml)",
-		Run: func(cmd *cobra.Command, args []string) {
-			// TODO: FIXME: config path
-			err := run_configuration.WriteConfiguration()
-			if err != nil {
-				common.ExitWithErr("Can`t save config: %+v", err)
-			}
-
-			fmt.Fprintf(os.Stdout, "Config created in ~/.config/kafeman/config.yaml\n")
 		},
 	}
